+++ packages/orq-rc/src/models/operations/filelist.ts
@@ -196,7 +196,7 @@ export const FileListData$inboundSchema: z.ZodType<
   file_name: z.string(),
   workspace_id: z.string(),
   created: z.string().datetime({ offset: true }).default(
-    "2025-11-05T05:24:58.805Z",
+    "2025-11-05T11:09:13.822Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -230,7 +230,7 @@ export const FileListData$outboundSchema: z.ZodType<
   bytes: z.number(),
   fileName: z.string(),
   workspaceId: z.string(),
-  created: z.date().default(() => new Date("2025-11-05T05:24:58.805Z"))
+  created: z.date().default(() => new Date("2025-11-05T11:09:13.822Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {